Continue reimplementing things.
Made the decision to remove gRPC from core functionality and instead allow it to be loaded as
a plugin. The plugin can be defined using the GAV of the gRPC executable without needing to
hardcode the classifier or type in this case, which decouples the synchronous gRPC solution
from other implementations like the reactive Salesforce implementation (see GH-47).

Also made the decision to further simplify imports by defaulting to scanning the classpath
for archives to extract. I've implemented this to run in parallel now to speed it up and
we can just look at dependencies provided in the POM dependencies block or as
tags on the plugin declaration itself, which will provide a cleaner solution to GH-38
going forwards.

The aforementioned functionality is not yet implemented but most of the scaffolding to
support this is now in place.