Configure unified artifact coordinate type.
This replaces the use of DependableCoordinate and ArtifactCoordinate
outside a couple of locations that depend on the artifact resolver
API as implementation detail. This will allow for adjusting the
internal implementation later much more easily going forwards.

The use of .setExtension on the artifact/dependency types used for
parameters is now deprecated and should be replaced with .setType
instead. This will be removed in v2.0.0.