GH-101: Use Aether artifact resolver directly
This removes the Maven Artifact Transfer dependency entirely and integrates
with the Maven Dependency Resolver API directly instead. This helps make this
more compatible with the future Maven 4.x API that will expect this usage and
this removes a set of dependencies that we do not actually need to be using.

As a small side effect, the project dependency resolution is now even
more optimised, as we no longer need to scan project dependencies again
to re-resolve them when the direct resolution mode is in use. This should
slightly improve build speeds in some more complicated edge cases and
configurations