GH-164: Fix bug where artifacts were being discarded under DIRECT resolution scope
This fixes the bug oriented around GH-164 where sources added as
dependencies were not being returned in the dependency paths when
global dependency resolution scope was set to TRANSITIVE. This
occurred due to the decision we make to only resolve dependencies
transitively if they lack the DIRECT DependencyResolutionScope
attribute on their descriptor directly. In the regular case, we'd
pass the initial resolved artifacts to the internal dependency
resolver which would return the initial artifacts AND their
transitive dependencies. If we were using DIRECT on a single
artifact and not globally, then this path would still be followed
but the dependency resolver would skip those DIRECT artifacts,
resulting in them not being included in the final paths.

Conversely if the dependency resolution globally was set to DIRECT
then this path is skipped and a more direct route is followed,
which avoids this issue.