GH-172: Fix dependency conflict resolution
This fixes the outstanding issue at GH-172 that creates an
issue when mixing project and plugin-based dependencies that
could be on different versions transitively. This only now
works because of GH-215 fixing the way dependencies get
resolved... we now resolve both project and plugin
dependencies at the exact same time, which allows for some
version resolution and conflict fixes to be performed
by Eclipse Aether.