GH-229: Do not fail if we have a partial dependency resolution failure
If we only have a partial dependency resolution failure, log the error
details and continue with the dependencies that we do have access to.
This prevents us having immediate issues with erroneous transitive
dependencies that we do not care about, letting Maven fail later
if the resources were actually required.

This change makes builds more resillient to poorly configured
dependencies, such as the JAXB dependency used in ehcache.