GH-394: Use Java argument files for command line arguments
This is far more reliable in cross-platform parsing consistency, and
mostly superceeds the use of Shlex, although we still rely on it
to deal with simple path quoting. We may be able to refactor this out
entirely in the future.

The change also helps support GH-397's implementation which introduces
a higher risk of bugs slipping in through the unreliable batch
specifications for argument parsing.