GH-438: Change unique name digest generation for archives to use full URIs
Currently we use absolute paths to generate unique names for archives
to avoid name collision, but since we now support resolution of archives
internally and recursively as part of optimisations for GH-438, there is
now a potential edge case where an archive within an archive can collide
with a different file tree by reusing the same name in a legal context.

Switching to using URIs to generate the digests reduces this risk, as
URIs encapsulate the full source of the file tree recursively.