GH-481: Fix bug where large builds crash on Windows
This change implements the argument file mechanism for protoc
command line flags, like we do already for Java invocations. This
helps avoid build failures on Windows when a large number of files
are being generated, since Windows has somewhat esoteric limits
on the command line length. Offloading this to a file works around
this constraint.