GH-481: Fix bug where large builds crash on Windows  (#482)
This change implements the argument file mechanism for protoc
command line flags, like we do already for Java invocations. This
helps avoid build failures on Windows when a large number of files
are being generated, since Windows has somewhat esoteric limits
on the command line length. Offloading this to a file works around
this constraint.

A reproduction of the issue for Windows is also included in the test
pack, which generates a large number of random protobuf sources on
each build and feeds them into the Maven plugin.

A fix for Mockito has also been included in this change to
squelch
warnings about agent attachments that I was encountering on Java 21
builds.

Fixes GH-481.