Improve Termux detection semantics
Termux detection now uses the java.vendor property
rather than a call to uname via a subprocess. This
will still catch most cases so that a warning can
be emitted without needing to have complicated
subprocess handling for this single edge case. This
also removes the warning for Android outside this
specific case as we do not test integration for this
scenario to know whether it is an issue or not.