GH-579: Resolve dependencies from MavenProject repository list
This replaces the use of a ProjectBuildingRequest to determine
the project remote repositories with inspection of the
current MavenProject instead.

It has been observed that ProjectBuildingRequest ignores any
repositories defined in the project POM, and only considers
the contenrs of the settings.xml in the user's ~/.m2.

Conversely, it was observed that the list within the
current MavenProject provides the correct expected list
when the user has overridden repositories in their POM.

This means that dependencies should now resolve correctly
in private repositories if the user overrides them per
project.

Fixes GH-579