GH-596: Restrict executor currency to avoid resource exhaustion
Large projects will currently attempt to extract a large
number of JARs in memory without an explicit concurrency
limit. This may result in resource exhaustion for CPU
and/or memory, affecting the stability of the plugin.

This reworks the ConcurrentExecutor to use a set of sensible
values that scales based on the available compute.