Generate table of contents in markdown documentation (#595)
This implements a JS hook that runs once the page has loaded and
generates a table of contents by sniffing out
`<h1>...<h6>` tags to construct a virtual header tree.

This allows us to make a table of contents without extending fluido or
hardcoding it in Markdown for the time being, which
should improve user experience with documentation.