v3.0.0 - improve how ordering of plugins works (#615)
Improve codegen ordering

This breaking change alters how we determine the order to run plugin
code generation in to be able to be relative to the builtin code
generators in protoc and the descriptor file generation in protoc.

Plugins now have an order of 0 rather than 100,000 by default. The
implication of setting a negative order is that it will run prior
to builtin and descriptor generation, and a positive order will run
afterwards.

The order of language generation is also now stable between builds.