Use a custom Plexus converter to defer URL parsing to the last minute
This will allow us to later register custom protocol handlers for
URLs correctly. We cannot do this at the moment because Plexus defaults
to trying to parse URLs using the URL default constructor, and that
only checks the system classloader for handlers that are registered
via the URLStreamHandlerProvider SPI. Our issue is that whilst this
would be fine in a normal application, Maven plugins utilise
ClassWorlds to separate classloading for each plugin. This means we
never run on the default classloader and anything we register with
our ServiceLoader will be ignored.

By defining this scaffolding, we can defer parsing URLs until we
enter the UrlResourceFetcher (now named UriResourceFetcher), and
we can inject our own logic and lookups procedurally.

This scaffold also allows us to pass Path and URI objects around
when working with Maven 3.8, which does not override the default
converters to include one for the NIO and URI APIs. This is
available from Maven 3.9 onwards, so we can eventually deprecate
and remove these workarounds.