Differentiate between artifacts and dependencies (#685)
Only dependencies make sense to carry exclusions and overridable
resolution depths. This makes no sense on plugins and can only result in
broken behaviour, so we remove the capability here.

This should never have been user-facing, and existing usages are
undocumented behaviour, so it should be safe to remove as part of this
bugfix.