GH-699: hash/digest verification of URL dependencies (#701)
This change introduces a new set of options that can be specified when
referencing protoc plugins by URL that allow users to specify the
expected digest of the resource to be downloaded. If the downloaded
resource does not match the digest, then the plugins are not executed,
and the build will fail with an error.

The aim is to allow users to verify that their dependencies have not
been tampered with prior to running anything. This is already performed
internally by Maven on Maven-based dependencies.

Digests will be able to be specified in the format
`md5:09f7e02f1290be211da707a266f153b3`,
`sha256:66a045b452102c59d840ec097d59d9467e13a3f34f6494e539ffd32c1bb35f18`,
etc for any supported JVM `MessageDigest` (this is usually a small set
including MD5, SHA-1, SHA-256, and SHA-512).

Users should consult the documentation for their Java version to see
which MessageDigest format are supported for their platform.

Users may _in theory_ be able to extend this by adding bouncy castle to
the classpath, although this will not be tested nor verified in this PR.

---

## TODO list

- [x] Implement Digest class for parsing, holding, comparing digests
- [x] Unit test Digest class
- [x] Implement Plexus DigestConverter to allow parsing digests to
Digest objects in pom.xml configuration blocks
- [x] Unit test DigestConverter class
- [x] Remove all references of `*.utils.Digests`, and replace with this
new `Digest` class
- [x] Add optional digest attribute to URL plugins
- [x] Validate digest of URL resource if the digest is provided in the
configuration
- [x] Add optional digest for protoc executable if specified as a URL.
- [x] Report user-friendly errors if digests do not match
- [x] Implement new integration tests
- [x] Minor version bump
- [x] Update documentation for MOJO goals
- [x] Update user guide markdown documentation

Closes GH-699.