GH-782: Implement ability to use sanctioned executable paths.
If you operate in an overly locked-down corporate environment that disallows running
shell/batch scripts or native executables outside sanctioned locations on your local
file system, you can specify the path here either via this configuration parameter
or via a property such that any executables are first moved to a directory within this
location. This is designed to be able to be used within a Maven profile if desired.