Implement mitigations for path length on windows (JDK-8315405)
At the time of writing, two OpenJDK bugs exist that cause us hassle on Windows:

- https://bugs.openjdk.org/browse/JDK-8315405
- https://bugs.openjdk.org/browse/JDK-8348664

On Windows, historical decisions mean that when invoking a native executable via
the CreateProcessW API
(https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-createprocessw),
the executable path cannot be longer than 260 characters.

For our hierarchy of generated directories to ensure build reproducibility, this
poses a visible issue, especially in our CI workflows (and as such, may impact other
users using a similar setup), since the path lengths when concatenated to the base
directory the project is cloned to will breach the 260 character limit.

A workaround is available as part of
https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation?tabs=registry
but this requires active system changes on Windows and possibly a set of changes
to OpenJDK itself, as discussed in JDK-8348664 above. At the time of writing, this
has not been implemented.

As a result, this commit changes how temporary spaces work to remove the nested
directory hierarchy that is provided by TemporarySpace APIs used internally for
build organization, and replaces them with a wrapper that names directories using
a fixed length SHA-256 digest of the previous components. This provides enough
"entropy" to avoid the risk of collisions in names, while providing a reproducible
file length until the problem is resolved. This should reduce the cases where this
issue arises until such a time that the problem is addressed properly, where we can
then revert this commit.