At the time of writing, two OpenJDK bugs exist that cause us hassle on Windows:
- https://bugs.openjdk.org/browse/JDK-8315405
- https://bugs.openjdk.org/browse/JDK-8348664

On Windows, historical decisions mean that when invoking a native executable via
the CreateProcessW API
(https://learn.microsoft.com/en-us/windows/win32/api/processthreadsapi/nf-processthreadsapi-createprocessw),
the executable path cannot be longer than 260 characters.

For our hierarchy of generated directories to ensure build reproducibility, this
poses a visible issue, especially in our CI workflows (and as such, may impact other
users using a similar setup), since the path lengths when concatenated to the base
directory the project is cloned to will breach the 260 character limit.

A workaround is available as part of
https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation?tabs=registry
but this requires active system changes on Windows and possibly a set of changes
to OpenJDK itself, as discussed in JDK-8348664 above. At the time of writing, this
has not been implemented.

For now, this is documented as a known issue.

[skip ci]