GH-802: Reimplement URL handling by decorating input streams with commons-compress
This reimplements all the low-level URL handling logic internally to remove
the SPI workaround that is currently in place, and use a recursive dispatcher
that has custom URLStreamHandlerFactories associated with it. These
implementations can delegate to archivers and compressors in Apache
commons-compress to provide support for protocols such as 'tar:gz:https://...'.

This effectively allows users to utilise arbitrary tarballs as the source of
their plugins now rather than just JARs.

Current implementation provides:
- Archivers for 'jar', 'war', 'ear', 'zip', 'tar'
- Compressors for 'gzip', 'bz2'

We can add further integrations such as 'xz', 'cpio', 'ar', 'ar2', 'lzma',
'7z', 'deflate', 'z', 'zstd', etc as needed by appending to the list
of providers in the new UrlFactory.

This has currently only been tested using tar.gz to prove that nested
retrieval works as expected. Other implementations will still need
unit tests to be added.

In addition, the tests for UriResourceFetcher are now invalid and need
a full rewrite to handle delegating to this new mechanism under the
hood.