GH-877: Implement Plexus converter that can convert algebraic sealed types
The implementation of this will expect users to take a 'kind' attribute
on the marker for a sealed type parameter which is used via reflection
to determine the concrete implementation class. Right now, we have to
provide this info on each implementation via an annotation such that
we can influence the behaviour that will eventually be provided for us
by GH-880.

This PR also changes the documentation and plugin interface so that users
will pass in a <plugins> list containing all plugins now rather than
having four distinct parameters for the same information. The exisiting
parameters have been kept and still work the same way, but are now
undocumented in the plugin guide, and are marked as deprecated-for-removal
in v5.0.0.

With this change, we will be able to use sealed implementation kinds for
the <protoc/> parameter in future changes, further supporting GH-877.